Modificação to tipo variáveis de localização

+++ Entities/Location.py
@@ -2,10 +2,10 @@ class Location:
     def __init__(self):
         pass
 
-    def set_x(self, x: float):
+    def set_x(self, x: int):
         self.__x = x
 
-    def set_y(self, y: float):
+    def set_y(self, y: int):
         self.__y = y
 
     def get_x(self):
